+++ examples/resources/harness_infrastructure_definition/import.sh
@@ -1,2 +1,2 @@
 # Import using the Harness application id, environment id, and infrastructure definition id
-terraform import harness_infrasturcture_definition.example <app_id>/<env_id>/<infradef_id>
+terraform import harness_infrastructure_definition.example <app_id>/<env_id>/<infradef_id>